make linking interactive

+++ link.sh
@@ -1,16 +1,19 @@
-ROOT="$(dirname "$(readlink -fm "$0")")"
-
-# config
-CONFIGS=( i3 sxhkd polybar kitty zsh picom kmonad )
-for c in "${CONFIGS[@]}"
-do
-	rm -rf $CONFIG/$c
-	ln -s $ROOT/$c $CONFIG/$c
-done
-
 # scripts
 rm -rf $HOME/.scripts
-ln -s $ROOT/scripts $HOME/.scripts
+ln -s $PWD/scripts $HOME/.scripts
 
 rm -rf $HOME/.profile
-ln -s $ROOT/profile $HOME/.profile
+ln -s $PWD/profile $HOME/.profile
+source $HOME/.profile
+
+# config
+CONFIGS=( i3 sxhkd polybar kitty zsh picom kmonad karabiner )
+for c in "${CONFIGS[@]}"
+do
+	echo "Link $c? (y/n)"
+	read RESPONSE
+	if [ $RESPONSE = 'y' ]; then
+		rm -rf $CONFIG/$c
+		ln -s $PWD/$c $CONFIG/$c
+	fi 
+done
